use double for velocity

+++ src/main/java/pdc/project/entity/EntityWithVelocity.java
@@ -1,8 +1,8 @@
 package pdc.project.entity;
 
 public interface EntityWithVelocity extends MoveableEntity {
-    int getVelocityX();
-    int getVelocityY();
-    void setVelocityX(int velocityX);
-    void setVelocityY(int velocityY);
+    double getVelocityX();
+    double getVelocityY();
+    void setVelocityX(double velocityX);
+    void setVelocityY(double velocityY);
 }
